delete

+++ index.php
@@ -26,7 +26,7 @@
         <td><?= $i; ?></td>
         <td>
             <a href="ubah.php">Ubah</a> |
-            <a href="hapus.php?id=<?= $produc['id']; ?>">Hapus</a>
+            <a href="hapus.php?id=<?= $produc['id']; ?>" onclick="return confirm('Apakah anda Yakin?')">Hapus</a>
         </td>
         <td><img src="img/<?= $produc['gambar']; ?> " width="100"></td>
         <td><?= $produc['nama']; ?></td>
